I've shoehorned a patch from Michael D�nzer that he wrote for quakeforge a long time ago that allows sound playback on machines that can't mmap /dev/dsp, like PowerPC machines.
The original patch was archived at:
http://www.geocrawler.com/mail/msg.php3?msg_id=4207733&list=856

The patch went in mostly smoothly, once I found where iD
kept their sound code.  Linux only at the moment, I haven't
bothered to look at snd_*.c in any of the other OS
directories.

It works for me, in that it doesn't interfere with mmapping
on my i386 machine, but I have no PPC machines to test that
it actually does what it's supposed to, so YMMV.