Fix server video loading
From h3xx:
There is a bug whereby, on some systems, when a command such as "map
*ntro.cin+base1" (as it is when starting Quake2 vanilla, defined in
default.cfg) is issued, the server gives the error that it can't find
"maps/ntro.inn.bsp" even though "video/ntro.cin" exists. After hacking at
this problem for a couple hours, I finally fixed it in two lines; the
problem lies in how the strings are used.

I am unsure what caused this problem to surface in one of my machines and
not the other (they both had exactly the same software), but this patch
fixes the problem on the one that was having it, and doesn't affect the one
that wasn't having it. Due to how I solved this problem, it would lead me
to believe that this was a simple programming mistake.